# packages/ai/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FLINT Verifiable AI Trust Layer",
        "version": "1.0.0",
        "description": "Verifiable AI Agent Infrastructure on Flare for DeFi Risk & Policy management."
    },
    "paths": {
        "/health": {
            "get": {
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "flare_ai_kit_integrated": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/consensus-decide": {
            "post": {
                "summary": "Multi-agent consensus decision",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "strategy_name": {
                                        "type": "string"
                                    },
                                    "portfolio": {
                                        "type": "object"
                                    },
                                    "market_data": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Decision with attestation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "decision_id": {
                                            "type": "string"
                                        },
                                        "decision": {
                                            "type": "object"
                                        },
                                        "attestation": {
                                            "type": "object"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/verify-decision/{decision_id}": {
            "get": {
                "summary": "Verify a decision by ID",
                "parameters": [
                    {
                        "name": "decision_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verification status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "decision_id": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}